refactor: move shared types to src
