+++ packages/components/src/Filter/components/FilterTriggerButtonRemovable/FilterTriggerButtonRemovable.tsx
@@ -2,7 +2,7 @@ import React, { forwardRef } from "react"
 import { Icon } from "@kaizen/component-library"
 import iconClear from "@kaizen/component-library/icons/clear.icon.svg"
 import { Tooltip } from "@kaizen/draft-tooltip"
-import { DataAttributes } from "../../types"
+import { DataAttributes } from "../../../types"
 import { isRefObject } from "../../utils/isRefObject"
 import {
   FilterTriggerButton,
