--- packages/components/src/Filter/types.ts
@@ -1,5 +1,0 @@
-export type OverrideClassName<T> = Omit<T, "className"> & {
-  classNameOverride?: string
-}
-
-export type DataAttributes = { [key: `data-${string}`]: string | undefined }
