+++ packages/components/src/types.ts
@@ -1,0 +1,4 @@
+export type OverrideClassName<T extends Record<string, any>> =
+  | Omit<T, "className"> & { classNameOverride?: string }
+
+export type DataAttributes = { [key: `data-${string}`]: string | undefined }
